Upd ocserv files

+++ openconnect/openconnect-src/ssl_update.sh
@@ -2,11 +2,25 @@
 
 set -e
 
-cd /opt/openconnect
+WORK_DIR="/opt/openconnect"
+CONTAINER_NAME="openconnect"
 
-docker-compose up certbot
+to_log () {
+    local text="$1"
+    echo "[$(date '+%F %T')] ${text}"
+}
+
+cd "$WORK_DIR" || exit 1
+
+if [[ -r ./docker-compose.yml ]]; then
+    to_log "Run certbot service container"
+    docker-compose up certbot
+    sleep 3
+    to_log "Reload ocserv config"
+    docker exec "$CONTAINER_NAME" occtl reload
+    to_log "Delete all unused docker images"
+    docker system prune -af
+fi
 
 # docker exec openconnect 'kill -HUP "$(pidof ocserv-main)"'
-docker exec openconnect 'occtl reload'
 
-docker system prune -af
